requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
python-docx
requests
